backend/requirements.txt
fastapi>=0.110.0
uvicorn[standard]>=0.28.0
python-multipart>=0.0.9
pymongo>=4.6.0
motor>=3.3.0
google-generativeai>=0.4.0
pillow>=10.3.0
python-dotenv>=1.0.0
pydantic>=2.6.0
aiohttp>=3.9.0
